refactor: split randomizer.rs into config + flag-key + orchestration
Break the 2160-line randomizer.rs into a randomizer/ module directory:
options (the Options struct, its enums, serde defaults), flag_key (the
base32 flag-key codec and Options encode/decode), and tests. mod.rs keeps
the randomize entry points and the randomize_inner pipeline.

Pure code movement: byte-for-byte identical ROM output across seeds and
the full flag matrix; all tests pass, clippy clean.
